024_OAI_Agent_Builder/Frontend/requirements.txt
flask
openai
python-dotenv
flask-cors
